+++ src/commands.rs
@@ -1,5 +1,6 @@
 pub mod filter;
 pub mod generate;
 pub mod lint;
+mod subsample;
 
-pub use self::{filter::filter, generate::generate, lint::lint};
+pub use self::{filter::filter, generate::generate, lint::lint, subsample::subsample};