+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+  * commands: Add subsample command.
+
 ## 0.7.1 - 2021-10-07
 
 ### Fixed